feat: Organizando Projeto Grupo 04
