+++ helper-api-dotnet-o5/Program.cs
@@ -1,3 +1,6 @@
+using helper_api_dotnet_o5.Services;
+using helper_api_dotnet_o5.Util.Interfaces;
+
 var builder = WebApplication.CreateBuilder(args);
 
 // Add services to the container.
@@ -6,6 +9,9 @@ builder.Services.AddControllers();
 // Learn more about configuring Swagger/OpenAPI at https://aka.ms/aspnetcore/swashbuckle
 builder.Services.AddEndpointsApiExplorer();
 builder.Services.AddSwaggerGen();
+
+//DI
+builder.Services.AddScoped<ISimulacaoInvestimentoService, SimulacaoInvestimentoService>();
 
 var app = builder.Build();
