+++ helper-api-dotnet-o5/Util/Class/Investimento.cs
@@ -1,0 +1,7 @@
+﻿namespace helper_api_dotnet_o5.Util.Class;
+
+public class Investimento
+{
+    public Decimal taxaAplicada { get; set; } = 0.0000M;
+    public Decimal jurosPago { get; set; } = 0.00M;
+}
