+++ helper-api-dotnet-o5/Util/DTOs/TaxaDTO.cs
@@ -1,0 +1,7 @@
+﻿namespace helper_api_dotnet_o5.Util.DTOs;
+
+public class TaxaDTO
+{
+    public required string data { get; set; }
+    public required string valor { get; set; }
+}
