+++ helper-api-dotnet-o5/Util/Enums/EPeriod.cs
@@ -1,0 +1,8 @@
+﻿namespace helper_api_dotnet_o5.Util.Enums;
+
+public enum EPeriod
+{
+    Days = 1,
+    Months = 2,
+    Years = 3
+}
